Add files via upload
Nodemcu_arduino_web.ino is the file for nodemcu from which the farm data is being sent to the web page and can be stored to used for future references. Nodemcu_arduino_web_for_arduino.ino file is for arduino through which various sensors are connected to compare and automatically on and off the motors and other apparatus in the farm. Also it include the code to interface it with NodeMCU to send the same data to the web page.